+++ packages/dashboard/src/components/editor/LocalEditor.tsx
@@ -6,7 +6,7 @@ import { RPC } from "@uimix/typed-rpc";
 import type {
   IRootToEditorRPCHandler,
   IEditorToRootRPCHandler,
-} from "@uimix/editor/src/state/IFrameRPC";
+} from "@uimix/editor/src/types/IFrameRPC";
 import { ProjectData } from "@uimix/model/src/collaborative";
 import { LoadingErrorOverlay } from "./LoadingErrorOverlay";
 import { DocumentMetadata, getDesktopAPI } from "../../types/DesktopAPI";
