+++ packages/editor/src/state/IFrameDataConnector.ts
@@ -3,7 +3,10 @@ import { parentWindowTarget } from "@uimix/typed-rpc/browser";
 import { RPC } from "@uimix/typed-rpc";
 import { ProjectState } from "./ProjectState";
 import { action } from "mobx";
-import { IEditorToRootRPCHandler, IRootToEditorRPCHandler } from "./IFrameRPC";
+import {
+  IEditorToRootRPCHandler,
+  IRootToEditorRPCHandler,
+} from "../types/IFrameRPC";
 import { throttle } from "lodash-es";
 import { ThumbnailTakerHost } from "./ThumbnailTakerHost";
 import { Clipboard } from "./Clipboard";
